Add Screen for manual move.
Manual moves works with return code controls, as the increment/decrement
does not work with long datatypes.

The VP for the moves is one per axis, and the return code encodes
the distance to travel in 0.01 mm steps (so 10 mm is 1000)
Homing is coded as "0". This allows that other designs encode different
steps at their pleasure.